✨ feat: Introduce MarkdownErrorBoundary for improved error handling in Markdown components
