+++ client/src/components/Chat/Messages/Content/Markdown.tsx
@@ -19,6 +19,7 @@ import { Citation, CompositeCitation, HighlightedText } from '~/components/Web/C
 import { Artifact, artifactPlugin } from '~/components/Artifacts/Artifact';
 import { langSubset, preprocessLaTeX, handleDoubleClick } from '~/utils';
 import CodeBlock from '~/components/Messages/Content/CodeBlock';
+import MarkdownErrorBoundary from './MarkdownErrorBoundary';
 import useHasAccess from '~/hooks/Roles/useHasAccess';
 import { unicodeCitation } from '~/components/Web';
 import { useFileDownload } from '~/data-provider';
@@ -172,44 +173,6 @@ type TContentProps = {
   content: string;
   isLatestMessage: boolean;
 };
-
-// Error Boundary Component
-interface ErrorBoundaryState {
-  hasError: boolean;
-  error?: Error;
-}
-
-class MarkdownErrorBoundary extends React.Component<
-  { children: React.ReactNode; content: string },
-  ErrorBoundaryState
-> {
-  constructor(props: { children: React.ReactNode; content: string }) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
-    return { hasError: true, error };
-  }
-
-  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
-    console.error('Markdown rendering error:', error, errorInfo);
-  }
-
-  componentDidUpdate(prevProps: { content: string }) {
-    if (prevProps.content !== this.props.content && this.state.hasError) {
-      this.setState({ hasError: false, error: undefined });
-    }
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return <p className="mb-2 whitespace-pre-wrap">{this.props.content}</p>;
-    }
-
-    return this.props.children;
-  }
-}
 
 const Markdown = memo(({ content = '', isLatestMessage }: TContentProps) => {
   const LaTeXParsing = useRecoilValue<boolean>(store.LaTeXParsing);
